gallyerimageversion: Fix occasional crash in create_update_resource
self.get_resource() returns a response object if the request didn't
raise an exception, otherwise it returns False. This causes an
occasional crash with "TypeError: 'bool' object is not subscriptable" in
the polling loop which expects the response to always be subscriptable.

While polling, if a request fails, just keep the old request and loop
again.